Update 1 (Part 2)
New:
Added a new config option in /config to auto generate a star spotify playlist. If this playlist is auto generated, the bot will add or remove songs from this playlist if they are starred and you have a spotify connection with waveform. This list will NOT contain EPs/LPs, but will contain singles or remixes. /stars is still available for you to view your local star list, which WILL have EPs/LPs (along with non-spotify songs)
NOTE: Because of how new this feature is and because of me not supporting it before, the auto generated playlist may contain some incorrect songs. Feel free to just alter these yourself!

Added /editdata ep-lp, which allows you to add and remove artists involved on an EP/LP, and rename EPs/LPs. (This will also allow you to edit tracklists on an EP/LP, in update 2!)

Changes:
Changed the name parser to handle EP/LPs formatted like this: Still Kids - EP.
Changed /setstar to add/remove starred songs to a spotify playlist if one is created for the user.

Fixes:
Fixed a series of typos in /sendmail when you send someone non-spotify mail.