Update 2 (Part 1)
New:
Added a message that is created whenever a song is added into hall of fame when being starred through /review or /setstar.

Changes:
Changed the star button on /review to add to the star spotify playlist, as it did not before.

Fixes:
Fixed a bug where /config wouldn't update the setup star playlist field to say "Setup." when the star playlist was setup.
Fixed a bug where stars on review titles would sometimes not appear or disappear properly when running /setstar.