Hotfix
New:
Added a setting in /config to set if you would like to be DM'd when being sent a new song in mailbox. This is by default turned on.

Fixes:
Fixed a bug where /getsong and /getep wouldn't display timestamps for reviews.
Potentially fixed a bug where /bugreport would crash the bot.
Fixed a bug where the auto art updater would crash the bot, making the bot crash on every review.