Update 3 (Part 2)
New:
Re-added /pushtoepreview, now setup as /pushtoepreview with_spotify and /pushtoepreview manually. This now works in exactly the same way as /review, so there should no longer be bugs and differences between the 2 commands.
Added /botstats, which displays statistics and information about the bot globally across all servers, as well as some general info about the database and userbase.
Added a new config to /config, Embed Color! This allows you to tweak the color of your embed across all commands to a specific hexcode you put in.

Changes:
Optimized every command that iterates through the database to run near instantly, instead of taking about 20-30 seconds to load.
Optimized the /config star spotify playlist feature to run in around half the time.
Removed the vocalist argument from every command, and changed display names instead display vocalists as artists, not separate with "ft.".
This was done to simplify the codebase, as it was a difficult feature to explain to those who didn't quite understand it, and caused a lot of specific bugs that were tough to fix.
Removed Least Favorite Song from /profile and replaced it with Favorite Artist.
Removed /topstars, /viewrating, and /serverconfig.
Removed the Average Overall Rating statistic from all commands that tracked it.
Removed the Number of 0s statistic from all commands that tracked it.
Removed the most reviewed and most starred artists statistics from /profile.
Removed the footer message that display the number of ratings given from /ratinglist.
Removed the average rating from the card view cards in /hall-of-fame.
The reason these were removed is due to me wanting to trim the stats shown to ensure that every command can run instantly with my (fairly flawed) system I made a long time ago. These could come back in some form, but not unless I do a lot of work to try and set things up to not make every command take a really long time to run.
Added spotify link hyperlinks to /getsong, /getreview, /starlist, /viewmail, /getep, and /hall-of-fame, if one exists.
Removed multiple entries from /serverstats to be in line with statistics changes, and moved some to /botstats.
Changed /epeditreview to be more consistent with /editreview, as well as making /editreview manually be able to review EPs/LPs on its own.

Renamings:
Renamed /stars to /starlist.

Fixes:
Fixed a massive amount of database issues and bad entries, and as a result, removed around 100-200 reviews from the database that were corrupted and/or broken. You are unlikely to notice a lot of these, but if you are missing a couple of reviews, that is why
Fixed a bug where /hall-of-fame would crash upon trying to use it.
Fixed a bug where /getsong would crash when trying to pull up songs from EPs/LPs.
Fixed a bug where /deletereview wouldn't remove a starred song from your star spotify playlist, if you delete a star review.
Fixed a bug where /getreview would crash when trying to pull up a song that isn't in the database.