Final Update (Part 4)
New:
Added 2 new moderation commands, /admindelreview and /adminbanuser, which allows me to delete any users review, or ban a user from reviewing.
Split up /getsong into 2 new subcommands: /getsong server and /getsong global.
/getsong server works exactly how /getsong currently works, but it only displays reviews of users in the server, rather than all users overall.
/getsong global displays a new summary list view, that does not display individual users reviews, instead displaying general information about the song, such as how many reviews it has, the review average, how many stars it has, etc.
The changes for /getsong were also applied to /getep.
Split up /setstar into 2 subcommands, /setstar ep and /setstar song, which handles setting stars for EPs/LPs and songs accordingly.

Changes:
Removed /epeditreview and instead replaced it with new subcommand groups for /editreview, which now has 4 different subcommands in total:
/editreview song with_spotify and /editreview song manually
/editreview ep with_spotify and /editreview ep manually
Removed the boolean argument for list view in /hall-of-fame and replaced it with subcommands for card view and list view.
Changed /getep to display the number of stars an EP/LP has.
Changed /nowplaying to display the server rating and the global rating of a song, as well as the global number of reviews and local number of reviews.
Removed the artist edit and song edit buttons from /review and /epreview when doing a spotify review.
Changed reviews with - to display as no text review reviews, to make it work with how it used to work.

Fixes:
Fixed some situations where remixes could be reviewed as a remix, even when the name formatting was messed up, adding artists that shouldn't exist into the database.
Fixed a bug where /getepreview would crash upon trying to pull up an EP/LP review, if you hadn't reviewed it before
Fixed a bug where /getep would display -1/10 as the EP/LP overall rating, if one was not given.
Fixed a bug where /getep and /getepreview would display stars when they shouldn't be displayed.
Fixed a bug where /getepreview would say the EP/LP review was not properly finished, when it was.
Fixed a bug where /editreview would crash if trying to add a rating to a no-rating review.