Hotfix
New:
Added a new command /reportsongdata, which will allow you report incorrect song data to me, for me to edit. This will NOT edit data in the database, it will just send me a notification to do so.

Changes:
Changed /nowplaying to better and more clearly display info (removed global # of reviews and global star count, and instead only show global rating average)

Fixes:
Fixed a bug where /getsong global would crash the bot entirely if it was ran on a song with 10 or more reviews.
Maybe fixed a bug where the bot would crash instead of giving a proper error log when trying to run a spotify command without playing something on spotify.
Fixed a bug where hall of fame would place in entries based off of global star counts, and not local server star counts.
Fixed a bug on /getsong server and /getep server which would not allow you to select the reviews of any page beyond page 1, on songs with more than 10 reviews.