Final Update (Part 5)
New:
Added /randomsong, which allows you to pull up a random song from the database that has been reviewed by someone, and gives you basic info akin to /nowplaying about it. (Does not pull up EPs/LPs)
Added /reportreview, which opens up a modal form to report a review made by a user.

Changes:
Changed /review manually to be able to automatically convert a song name input as Song (Artist Remix) into the proper formatting.
Changed /editreview to update the timestamp when a review is edited, so an edited review will show the correct date of when the review was last changed when pulled up.
Changed the song name parser to handle songs like Artist 1 & Artist 2 - Song (Artist 1 & Artist 2 Remix). It treats these as not a remix, as it is impossible to tell who is the original artist from the limited info the bot has.
Renamed /clearmail to /clearallmail.
Made it so that animated avatars are properly disabled when a user pfp is displayed on most commands.
Made it so the bot now creates profiles for all users in a server when joining it.

Fixes:
Fixed a bug where /getep server would display global averages and global data for the songs on an EP/LP
Fixed a bug where /epreview manually wouldn't enforce tracklist checks on manually reviewed songs.
Fixed a bug where /nowplaying wouldn't properly display if you had a song starred.