Waveform Version 1.2 (Major Update)
New:
Added a new subcommand to /review, /review spotify_link!
This subcommand allows you to review singles by passing in a spotify link. This will only work with open.spotify.com links! Will not work with spotify.link links.
The process for reviewing is the same otherwise, you just use a spotify link to review rather than typing in artist names manually or playing a song on spotify! Does not require you to be logged in either.
/epreview spotify_link will be added in the next update!

Changes:
Renamed Stars to Favorites, renamed it across all commands, and removed the 7/10 or higher requirement.
This was done to allow for your favorites list to be curated exactly how you want, rather than only songs you rate highly. Beyond the name change and removal of 7/10 or higher requirement, they work exactly the same.
Changed /viewmail to have a user argument, allowing you to view the mailbox list of other users in a server you are in.

Fixes:
Fixed a bug where the bot would crash upon trying to send spotify.link links. Waveform now gives a proper error message saying these links are not supported.
As of present, these links will not be supported as the API doesn't react nicely with them. I may change this in the future if I figure it out!
Fixed a bug where mailbox songs would not work properly with Waveform Mailbox if run with /review manually or /epreview manually.
Potentially fixed a bug where EPs/LPs wouldn't properly work with Waveform Mailbox features.