Waveform Version 1.21 (Minor Update/Hotfix)
Changes:
Changed /viewmail and /deletemail to be usable in DMs with Waveform.
/viewmail will display your mailbox across ALL servers when pulled up via DM. This is the only way to view your cross-server mailbox besides the spotify playlist itself.
/nowplaying will now display if the song you are listening to is in your mailbox, and who sent you the song.
This will not work with EPs/LPs, as they are not pulled up with /nowplaying.

Fixes:
Fixed a bug where /sendmail would crash the bot when ran without a link argument.
Fixed a bug where running /getreview on a mailbox review sent by a user who left the server would result in a crash.
Fixed a bug where albums could not be properly manually removed by manually typing in info into /deletemail.
Fixed a bug where discord usernames were being used for display names over the actual "Display Name"