Waveform Version 1.3 (Major Update)
New:
Added support for using /epreview on 1 track albums with the 1 track being over 20 minutes, such as Fishmans - LONG SEASON.
Added /epreview spotify_link.
This works basically the same as /epreview manually, but autofilling in the artist and EP/LP names. It currently does not give you a tracklist like /epreview with_spotify does, but this will be added in a future update.

Changes:
/nowplaying now displays the EP/LP a song is from, alongside who sent you the song via mailbox, in a mixed bottom embed message.
/nowplaying now will always display the EP/LP a song is from, even if its not in the database, as long as its connected to an EP/LP on spotify.

Fixes:
Fixed a bug where the "from EP/LP" footer message was bugged on /nowplaying, and would never appear in the right conditions.
Did some slight code optimizations to /nowplaying to make a bit more consistent.