Waveform Version 1.4 (Major Update)
New:
Added a "Reviews" button to /nowplaying and /randomsong, that allows you to view the reviews of that song for your server, just by clicking the button!

Changes:
Changed /epreview spotify_link to automatically fill in the tracklist of the EP/LP based on the spotify tracklist.
Changed the dropdown menu of /getep and /getsong to display the users rating of a song/EP/LP drop down list, to make it easier to see what a users review is when selecting someone on the list.
/ratingcount has been renamed to /allratings.
Added song hyperlinks into /getartist, and generally spruced up the command to look a lot cleaner and easier to read.

Fixes:
Fixed a bug where formatting would get horribly messed up on some /getartist pages.
Fixed some various random typos across some commands.
Fixed a bug where /nowplaying would display "EP EP" on some songs that are in an EP/LP