Waveform Version 1.6 (Major Update)
New:
Added more last.fm support:
/getep will now display overall album scrobbles for each reviewer as well as yourself. It works identically to /getsong.
/nowplaying will now use Last.fm data if Spotify data is not detected or you are not logged into Spotify.
/nowplaying now displays a link to the last.fm page, if it can get one.
Added a config to /config to show/hide your scrobble count.
Added a new command /playing, which displays songs people in a discord server Waveform is in are playing (on Spotify, if logged into Spotify on Waveform)
This is just the first iteration of this command. I plan to add a lot more to it later, such as scrobble data, and users who are listening to music on last.fm but not on Spotify!

Changes:
Removed the scrobble detection system from EP/LP cases with /sendmail, as it was buggy and didn't make much logical sense.
Changed /config to change its embed color based on the new embed color you put in.
Changed /config to have the ✅ and ❌ emojis for "true" or "false", as they look nicer.
Changed Spotify Web API errors to be caught and display a better error message.
Removed some footer text from /serverstats about the hall of fame command.

Fixes:
Fixed a bug where /config would crash when running the embed color change button twice.
Fixed a crash with /getep when a single was playing.