Waveform Version 1.61 (Minor Update)
New:
/playing now works with Last.fm playback. If you are a Spotify user, it will prioritize the spotify playback. If Spotify playback can't be found, it will check for last.fm playback.
Added a scrobble counter to /playing, if a user is logged into Last.fm.
Added an emoji indicator next to the name of the user listening to a song on /playing, to mark if they're listening through Spotify or Last.fm.

Changes:
Changed the song name on /playing to be a spotify hyperlink to the song.

Fixes:
Fixed a bug where the show_songs argument on /getepreview would have the opposite effect
Fixed a bug where /sendmail would fail if a last.fm user play count couldn't be found (displaying as undefined)
Fixed a bug where /playing would crash if no one was playing a song.