Waveform Version 1.8 (Major Update)
New:
Added the ability to disable the rating system serverwide with /serverconfig. This will still allow reviews, but all rating specific commands will be disabled, and commands that have ratings and reviews will be treated as if those ratings don't exist.
This is mostly for people who want to use the bot, but don't want to see /10 ratings on songs. They can instead just use it for reviews, or for last.fm related features and mailbox related features.
Only server admins can edit this config.
Added a new command /whoknows, which will show the scrobble counts of an album, song, or artist from all users logged into waveform in a server.
This works similarly to the commands from the bot Chuu, if you are familiar with that!
Added a new filter to /viewmail, filter_user, which allows you to view your mailbox sends from a specific user!
Added alphabetical sort modes to /viewreviews.

Changes:
Renamed all instances of "scrobbles" to "plays", as its a little less confusing.
Renamed the user argument in /viewmail to mailbox_user.
Changed /setupmailbox text to be more clear.
Adjusted a lot of wording to better handle the rating disable system, or songs without any ratings.

Fixes:
Fixed a typo in /botstats.