Waveform Version 1.8.4 (Hotfix)
Changes:
Changed /login to use ephemeral messages instead of straight up messages.
Changed /profile to not show the 10/10 stat if Disable Ratings is turned on.
Changed /whoknows to display the song name that it couldn't find.
Changed /viewreviews to not display Rating: N/A if Disable Ratings is turned on.
Changed /getalbum and /getalbumreview related commands to automatically append LP at the end of an album, if LP or EP is not specified.

Fixes:
Fixed a bug where a Spotify Web API error would occur upon reviewing a song with a massive remixer or artist list
Fixed a bug where /whoknows would display lots of undefined if it couldn't find a song.
Fixed a bug where Last.fm and Spotify URLs would attempt to display even if they displayed as undefined in /nowplaying.