Waveform Version 1.8.7 (Hotfix)
Changes:
Changed /viewmail to be /viewmail spotify, to make it clearer what it is.
Changed /review to display an ephemeral message informing the user their review was not submitted, if it gets automatically deleted from inactivity.

Fixes:
Fixed a bug where /editreview wouldn't work in some cases.
ACTUALLY fixed a hard crash this time when running /viewmail in a DM.
Fixed a bug where /sendmail would not display proper error messages.
Fixed a bug where /review would not auto delete after inactivity if any of the edit buttons were pressed but the confirm review button was not pressed.