Waveform Version 1.8.8 (Hotfix)
Changes:
Split up /deletereview into the following subcommands:
/deletereview song with_spotify
/deletereview song manually
/deletereview album with_spotify
/deletereview album manually
This helps make it consistent with /editreview. It works exactly the same.
Removed the embed preview from /editreview, as it was inconsistent and kind of buggy, and looked bad.
Changed the /editreview return text to show the new review in a nicer way.

Fixes:
Fixed a bug where /review reviews would time out after 10 seconds.
Fixed an error when running /deletereview on a song/EP/LP that hasn't been reviewed in the database before.
Fixed a hard error when pulling up certain songs on /getep server.