Waveform Version 1.8.10 (Minor Update)
New:
Added new fields to /albumreview, least_fav_songs and fav_songs, which allow you to specify a favorite and least favorite songs on an album review. These can be typed any way you wish, and can be omitted if you don't need them.
The ability to edit this will come in a later update. For now, if you want to edit, you will have to re-review the EP/LP.
These do not show up if you review each individual song, as it's already obvious what are your favorites/not favorites through that.

Changes:
Made /lfmtop look a bit nicer.

Fixes:
Fixed a crash with /albumreview manually.